Add make-bitwise-generator after all
As it turns out, a SRFI-121 generator has no special disjoint type. It just has
to conform to the calling convention described in that specification. The spec
explicitly says: “They are just procedures that conform to a calling conven-
tion, so you can construct a generator with lambda.”

And that's what this does.

Adjusts the section in README about this. Also ships with tests, based on the
example in SRFI-151.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>